+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - `domain::hfbi::{RisultatoHFBI, ValoriIntermediHFBI}`
 - Add `impl From<(f32, &AreaNISECI)>` to `domain::niseci::StatoEcologicoNISECI`
 - Add `impl From<f32>` to `domain::hfbi::StatoEcologicoHFBI`
+- Add `impl Default` to `domain::niseci::{ClassiEtaSpecieNISECI, ClassiEtaAlieniNISECI, InfoPopolazioniNISECI}`
 
 ### Changed